Developments up to question number 4.

+++ src/main/java/com/example/restservice/model/House.java
@@ -2,15 +2,15 @@ package com.example.restservice.model;
 
 public class House {
 	
-	private final long houseid;
+	private final int houseid;
 	private final String housetypeid;
 
-	public House(long houseid, String housetypeid) {
+	public House(int houseid, String housetypeid) {
 		this.houseid = houseid;
 		this.housetypeid = housetypeid;
 	}
 
-	public long getHouseid() {
+	public int getHouseid() {
 		return houseid;
 	}
